# PERF: ffill/bfill with non-numpy dtypes